\name{Skillings.Mack}
\alias{Skillings.Mack}
\docType{package}
\title{The Skillings-Mack test Statistic for block designs with missing obseravtions
}
\description{
Implements a nonparametric statistical test for data from block designs with missing observations. The Skillings-Mack test statistic is a generalization of the statistic used in Friedman's ANOVA method and in Durbin's rank test.
}
\details{
\tabular{ll}{
Package: \tab Skillings.Mack\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-03-28\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Patchanok Srisuradetchai, John J. Borkowski \cr
Maintainer: Patchanok Srisuradetchai <s.patchanok@gmail.com>
}
\references{
Skillings, J.H., Mack, G.A.: On the Use of a Friedman-Type Statistic in Balanced and Unbalanced Block Designs. Technometrics 23(2), 171-177 (1981)
}
